Cast attention to float32 before scoring to avoid fp16 epsilon underflow
The 1e-8 epsilon in attn2score underflows to exactly 0 in float16, whose
smallest subnormal is ~6e-8. A head whose attention falls entirely outside
both the instruction and data ranges then divides 0/0, giving NaN, and
np.mean propagates it to the whole score.

Attention-Tracker's process_attn casts to float32 before the same
arithmetic (detector/utils.py:7). This matches that.

Reproduced on Qwen2-1.5B fp16 with a config listing all 336 head positions:
layer 0 heads 3 and 9 and layer 16 head 1 are attention-sink heads with zero
mass in both ranges. Scores were NaN; with the cast, 0.536 clean and 0.388
injected. The shipped 14-head config is unaffected (0.920 / 0.541 before and
after).

+++ vllm_hook_plugins/vllm_hook_plugins/analyzers/attention_tracker_analyzer.py
@@ -92,7 +92,7 @@ class AttntrackerAnalyzer:
         for attention, input_range in zip(batch_attention, batch_input_range):
             scores = []
             for _, layer_data in attention.items():
-                attn_np = layer_data['attention'].numpy()  # [num_heads, seq_len] — single transfer
+                attn_np = layer_data['attention'].to(torch.float32).numpy()  # [num_heads, seq_len] — single transfer
 
                 inst_attn = attn_np[:, input_range[0][0]:input_range[0][1]]  # [num_heads, inst_len]
                 data_attn = attn_np[:, input_range[1][0]:input_range[1][1]]  # [num_heads, data_len]
